feat: add modules in project
