+++ app.js
@@ -1,0 +1,1 @@
+import {keyboardEn, keyboardRu, keyboardEnShift, keyboardRuShift} from './components/layots.js';
