+++ components/button-block-handler.js
@@ -1,0 +1,5 @@
+import {changeTextareaValue} from './change-textarea-value.js';
+
+export function btnBlockHandler(event) {
+  changeTextareaValue(event.target.textContent);
+}
