+++ components/layots.js
@@ -1,0 +1,7 @@
+export const keyboardEn = [{ symbol: '`', keyCode: 'backquote' }, { symbol: '1', keyCode: 'digit1' }, { symbol: '2', keyCode: 'digit2' }, { symbol: '3', keyCode: 'digit3' }, { symbol: '4', keyCode: 'digit4' }, { symbol: '5', keyCode: 'digit5' }, { symbol: '6', keyCode: 'digit6' }, { symbol: '7', keyCode: 'digit7' }, { symbol: '8', keyCode: 'digit8' }, { symbol: '9', keyCode: 'digit9' }, { symbol: '0', keyCode: 'digit0' }, {symbol: '-', keyCode: 'minus' }, {symbol: '=', keyCode: 'equal' }, { symbol: 'BACKSPACE', keyCode: 'backspace' }, { symbol: 'TAB', keyCode: 'tab' }, { symbol: 'q', keyCode: 'keyq' }, { symbol: 'w', keyCode: 'keyw' }, { symbol: 'e', keyCode: 'keye' }, { symbol: 'r', keyCode: 'keyr' }, { symbol: 't', keyCode: 'keyt' }, { symbol: 'y', keyCode: 'keyy' }, { symbol: 'u', keyCode: 'keyu' }, { symbol: 'i', keyCode: 'keyi' }, { symbol: 'o', keyCode: 'keyo' }, { symbol: 'p', keyCode: 'keyp' }, { symbol: '[', keyCode: 'bracketleft' }, { symbol: ']', keyCode: 'bracketright' }, { symbol: '\\',  keyCode: 'backslash' }, { symbol: 'DEL', keyCode: 'delete' }, { symbol: 'CAPS LOCK', keyCode: 'capslock' }, { symbol: 'a', keyCode: 'keya' }, { symbol: 's', keyCode: 'keys' }, {  symbol: 'd', keyCode: 'keyd' }, { symbol: 'f', keyCode: 'keyf' }, { symbol: 'g', keyCode: 'keyg' }, { symbol: 'h', keyCode: 'keyh' }, { symbol: 'j', keyCode: 'keyj' }, { symbol: 'k', keyCode: 'keyk' }, { symbol: 'l', keyCode: 'keyl' }, { symbol: ';', keyCode: 'semicolon' }, { symbol: '\'', keyCode: 'quote' }, { symbol: 'ENTER', keyCode: 'enter' }, { symbol: 'SHIFT', keyCode: 'shiftleft' }, { symbol: 'z', keyCode: 'keyz' }, { symbol: 'x', keyCode: 'keyx' }, { symbol: 'c', keyCode: 'keyc' }, { symbol: 'v', keyCode: 'keyv' }, { symbol: 'b', keyCode: 'keyb' }, { symbol: 'n', keyCode: 'keyn' }, { symbol: 'm', keyCode: 'keym' }, { symbol: ',', keyCode: 'comma' }, { symbol: '.', keyCode: 'period' }, { symbol: '/', keyCode: 'slash' }, { symbol: '▲', keyCode: 'arrowup' }, { symbol: 'SHIFT', keyCode: 'shiftright' }, { symbol: 'CTRL', keyCode: 'controlleft' }, { symbol: 'WIN', keyCode: 'metaleft' }, { symbol: 'ALT', keyCode: 'altleft' }, { symbol: ' ', keyCode: 'space' }, { symbol: 'ALT', keyCode: 'altright' }, { symbol: '◄', keyCode: 'arrowleft' }, { symbol: '▼', keyCode: 'arrowdown' }, { symbol: '►', keyCode: 'arrowright' }, { symbol: 'CTRL', keyCode: 'controlright' }, ];
+
+export const keyboardRu = [{ symbol: 'ё', keyCode: 'backquote' }, { symbol: '1', keyCode: 'digit1' }, { symbol: '2', keyCode: 'digit2' }, { symbol: '3', keyCode: 'digit3' }, { symbol: '4', keyCode: 'digit4' }, { symbol: '5', keyCode: 'digit5' }, { symbol: '6', keyCode: 'digit6' }, { symbol: '7', keyCode: 'digit7' }, { symbol: '8', keyCode: 'digit8' }, { symbol: '9', keyCode: 'digit9' }, { symbol: '0', keyCode: 'digit0' }, {symbol: '-', keyCode: 'minus' }, {symbol: '=', keyCode: 'equal' }, { symbol: 'BACKSPACE', keyCode: 'backspace' }, { symbol: 'TAB', keyCode: 'tab' }, { symbol: 'й', keyCode: 'keyq' }, { symbol: 'ц', keyCode: 'keyw' }, { symbol: 'у', keyCode: 'keye' }, { symbol: 'к', keyCode: 'keyr' }, { symbol: 'е', keyCode: 'keyt' }, { symbol: 'н', keyCode: 'keyy' }, { symbol: 'г', keyCode: 'keyu' }, { symbol: 'ш', keyCode: 'keyi' }, { symbol: 'щ', keyCode: 'keyo' }, { symbol: 'з', keyCode: 'keyp' }, { symbol: 'х', keyCode: 'bracketleft' }, { symbol: 'ъ', keyCode: 'bracketright' }, { symbol: '\\',  keyCode: 'backslash' }, { symbol: 'DEL', keyCode: 'delete' }, { symbol: 'CAPS LOCK', keyCode: 'capslock' }, { symbol: 'ф', keyCode: 'keya' }, { symbol: 'ы', keyCode: 'keys' }, {  symbol: 'в', keyCode: 'keyd' }, { symbol: 'а', keyCode: 'keyf' }, { symbol: 'п', keyCode: 'keyg' }, { symbol: 'р', keyCode: 'keyh' }, { symbol: 'о', keyCode: 'keyj' }, { symbol: 'л', keyCode: 'keyk' }, { symbol: 'д', keyCode: 'keyl' }, { symbol: 'ж', keyCode: 'semicolon' }, { symbol: 'э', keyCode: 'quote' }, { symbol: 'ENTER', keyCode: 'enter' }, { symbol: 'SHIFT', keyCode: 'shiftleft' }, { symbol: 'я', keyCode: 'keyz' }, { symbol: 'ч', keyCode: 'keyx' }, { symbol: 'с', keyCode: 'keyc' }, { symbol: 'м', keyCode: 'keyv' }, { symbol: 'и', keyCode: 'keyb' }, { symbol: 'т', keyCode: 'keyn' }, { symbol: 'ь', keyCode: 'keym' }, { symbol: 'б', keyCode: 'comma' }, { symbol: 'ю', keyCode: 'period' }, { symbol: '.', keyCode: 'slash' }, { symbol: '▲', keyCode: 'arrowup' }, { symbol: 'SHIFT', keyCode: 'shiftright' }, { symbol: 'CTRL', keyCode: 'controlleft' }, { symbol: 'WIN', keyCode: 'metaleft' }, { symbol: 'ALT', keyCode: 'altleft' }, { symbol: ' ', keyCode: 'space' }, { symbol: 'ALT', keyCode: 'altright' }, { symbol: '◄', keyCode: 'arrowleft' }, { symbol: '▼', keyCode: 'arrowdown' }, { symbol: '►', keyCode: 'arrowright' }, { symbol: 'CTRL', keyCode: 'controlright' }, ];
+
+export const keyboardEnShift = [{ symbol: '~', keyCode: 'backquote' }, { symbol: '!', keyCode: 'digit1' }, { symbol: '@', keyCode: 'digit2' }, { symbol: '#', keyCode: 'digit3' }, { symbol: '$', keyCode: 'digit4' }, { symbol: '%', keyCode: 'digit5' }, { symbol: '^', keyCode: 'digit6' }, { symbol: '&', keyCode: 'digit7' }, { symbol: '*', keyCode: 'digit8' }, { symbol: '(', keyCode: 'digit9' }, { symbol: ')', keyCode: 'digit0' }, {symbol: '_', keyCode: 'minus' }, {symbol: '+', keyCode: 'equal' }, { symbol: 'BACKSPACE', keyCode: 'backspace' }, { symbol: 'TAB', keyCode: 'tab' }, { symbol: 'q', keyCode: 'keyq' }, { symbol: 'w', keyCode: 'keyw' }, { symbol: 'e', keyCode: 'keye' }, { symbol: 'r', keyCode: 'keyr' }, { symbol: 't', keyCode: 'keyt' }, { symbol: 'y', keyCode: 'keyy' }, { symbol: 'u', keyCode: 'keyu' }, { symbol: 'i', keyCode: 'keyi' }, { symbol: 'o', keyCode: 'keyo' }, { symbol: 'p', keyCode: 'keyp' }, { symbol: '{', keyCode: 'bracketleft' }, { symbol: '}', keyCode: 'bracketright' }, { symbol: '|',  keyCode: 'backslash' }, { symbol: 'DEL', keyCode: 'delete' }, { symbol: 'CAPS LOCK', keyCode: 'capslock' }, { symbol: 'a', keyCode: 'keya' }, { symbol: 's', keyCode: 'keys' }, {  symbol: 'd', keyCode: 'keyd' }, { symbol: 'f', keyCode: 'keyf' }, { symbol: 'g', keyCode: 'keyg' }, { symbol: 'h', keyCode: 'keyh' }, { symbol: 'j', keyCode: 'keyj' }, { symbol: 'k', keyCode: 'keyk' }, { symbol: 'l', keyCode: 'keyl' }, { symbol: ':', keyCode: 'semicolon' }, { symbol: '"', keyCode: 'quote' }, { symbol: 'ENTER', keyCode: 'enter' }, { symbol: 'SHIFT', keyCode: 'shiftleft' }, { symbol: 'z', keyCode: 'keyz' }, { symbol: 'x', keyCode: 'keyx' }, { symbol: 'c', keyCode: 'keyc' }, { symbol: 'v', keyCode: 'keyv' }, { symbol: 'b', keyCode: 'keyb' }, { symbol: 'n', keyCode: 'keyn' }, { symbol: 'm', keyCode: 'keym' }, { symbol: '<', keyCode: 'comma' }, { symbol: '>', keyCode: 'period' }, { symbol: '?', keyCode: 'slash' }, { symbol: '▲', keyCode: 'arrowup' }, { symbol: 'SHIFT', keyCode: 'shiftright' }, { symbol: 'CTRL', keyCode: 'controlleft' }, { symbol: 'WIN', keyCode: 'metaleft' }, { symbol: 'ALT', keyCode: 'altleft' }, { symbol: ' ', keyCode: 'space' }, { symbol: 'ALT', keyCode: 'altright' }, { symbol: '◄', keyCode: 'arrowleft' }, { symbol: '▼', keyCode: 'arrowdown' }, { symbol: '►', keyCode: 'arrowright' }, { symbol: 'CTRL', keyCode: 'controlright' }, ];
+
+export const keyboardRuShift = [{ symbol: 'ё', keyCode: 'backquote' }, { symbol: '!', keyCode: 'digit1' }, { symbol: '"', keyCode: 'digit2' }, { symbol: '№', keyCode: 'digit3' }, { symbol: ';', keyCode: 'digit4' }, { symbol: '%', keyCode: 'digit5' }, { symbol: ':', keyCode: 'digit6' }, { symbol: '?', keyCode: 'digit7' }, { symbol: '*', keyCode: 'digit8' }, { symbol: '(', keyCode: 'digit9' }, { symbol: ')', keyCode: 'digit0' }, {symbol: '_', keyCode: 'minus' }, {symbol: '+', keyCode: 'equal' }, { symbol: 'BACKSPACE', keyCode: 'backspace' }, { symbol: 'TAB', keyCode: 'tab' }, { symbol: 'й', keyCode: 'keyq' }, { symbol: 'ц', keyCode: 'keyw' }, { symbol: 'у', keyCode: 'keye' }, { symbol: 'к', keyCode: 'keyr' }, { symbol: 'е', keyCode: 'keyt' }, { symbol: 'н', keyCode: 'keyy' }, { symbol: 'г', keyCode: 'keyu' }, { symbol: 'ш', keyCode: 'keyi' }, { symbol: 'щ', keyCode: 'keyo' }, { symbol: 'з', keyCode: 'keyp' }, { symbol: 'х', keyCode: 'bracketleft' }, { symbol: 'ъ', keyCode: 'bracketright' }, { symbol: '/',  keyCode: 'backslash' }, { symbol: 'DEL', keyCode: 'delete' }, { symbol: 'CAPS LOCK', keyCode: 'capslock' }, { symbol: 'ф', keyCode: 'keya' }, { symbol: 'ы', keyCode: 'keys' }, {  symbol: 'в', keyCode: 'keyd' }, { symbol: 'а', keyCode: 'keyf' }, { symbol: 'п', keyCode: 'keyg' }, { symbol: 'р', keyCode: 'keyh' }, { symbol: 'о', keyCode: 'keyj' }, { symbol: 'л', keyCode: 'keyk' }, { symbol: 'д', keyCode: 'keyl' }, { symbol: 'ж', keyCode: 'semicolon' }, { symbol: 'э', keyCode: 'quote' }, { symbol: 'ENTER', keyCode: 'enter' }, { symbol: 'SHIFT', keyCode: 'shiftleft' }, { symbol: 'я', keyCode: 'keyz' }, { symbol: 'ч', keyCode: 'keyx' }, { symbol: 'с', keyCode: 'keyc' }, { symbol: 'м', keyCode: 'keyv' }, { symbol: 'и', keyCode: 'keyb' }, { symbol: 'т', keyCode: 'keyn' }, { symbol: 'ь', keyCode: 'keym' }, { symbol: 'б', keyCode: 'comma' }, { symbol: 'ю', keyCode: 'period' }, { symbol: ',', keyCode: 'slash' }, { symbol: '▲', keyCode: 'arrowup' }, { symbol: 'SHIFT', keyCode: 'shiftright' }, { symbol: 'CTRL', keyCode: 'controlleft' }, { symbol: 'WIN', keyCode: 'metaleft' }, { symbol: 'ALT', keyCode: 'altleft' }, { symbol: ' ', keyCode: 'space' }, { symbol: 'ALT', keyCode: 'altright' }, { symbol: '◄', keyCode: 'arrowleft' }, { symbol: '▼', keyCode: 'arrowdown' }, { symbol: '►', keyCode: 'arrowright' }, { symbol: 'CTRL', keyCode: 'controlright' }, ];
