+++ index.html
@@ -12,6 +12,6 @@
     <title>Virtual Keyboard</title>
   </head>
   <body>
-    <script src="./index.js"></script>
+    <script src="./index.js" type="module"></script>
   </body>
 </html>
